Implements instance pooling.
However, benchmarking shows that this is actually slower than the
current implementation. There may be a bug in the pooling implementation
that hasn't been identified or we may simply be over estimating the
performance impact that creating an instance has. I'm going to push this
branch to a private parking lot fork for posterity, but I don't think it
makes sense to merge this.